refactor: initial commit for refactoring to DDD
