+++ utils/constants/index.js
@@ -1,0 +1,11 @@
+const MAX_DEPOSIT_PLANS = 2;
+const DEPOSIT_TYPE_ONE_TIME = "One time";
+const DEPOSIT_TYPE_MONTHLY = "Monthly";
+const VALID_PLAN_TYPES = [DEPOSIT_TYPE_ONE_TIME, DEPOSIT_TYPE_MONTHLY];
+
+module.exports = {
+  MAX_DEPOSIT_PLANS,
+  DEPOSIT_TYPE_ONE_TIME,
+  DEPOSIT_TYPE_MONTHLY,
+  VALID_PLAN_TYPES,
+};
